Refactor control runtime into focused modules
